+++ src/modules/computePlans/ComputePlansApi.ts
@@ -32,42 +32,46 @@ export const retrieveComputePlan = (key: string): AxiosPromise<ComputePlanT> =>
 
 export const listComputePlanTesttuples = (
     key: string,
+    searchFilters: SearchFilterType[],
     page: number,
     pageSize?: number
 ): AxiosPromise<PaginatedApiResponse<TesttupleT>> => {
     return API.authenticatedGet(
         URLS.LIST_TESTTUPLES.replace('__KEY__', key),
-        getApiOptions([], page, pageSize)
+        getApiOptions(searchFilters, page, pageSize)
     );
 };
 export const listComputePlanTraintuples = (
     key: string,
+    searchFilters: SearchFilterType[],
     page: number,
     pageSize?: number
 ): AxiosPromise<PaginatedApiResponse<TraintupleT>> => {
     return API.authenticatedGet(
         URLS.LIST_TRAINTUPLES.replace('__KEY__', key),
-        getApiOptions([], page, pageSize)
+        getApiOptions(searchFilters, page, pageSize)
     );
 };
 export const listComputePlanCompositeTraintuples = (
     key: string,
+    searchFilters: SearchFilterType[],
     page: number,
     pageSize?: number
 ): AxiosPromise<PaginatedApiResponse<CompositeTraintupleT>> => {
     return API.authenticatedGet(
         URLS.LIST_COMPOSITE_TRAINTUPLES.replace('__KEY__', key),
-        getApiOptions([], page, pageSize)
+        getApiOptions(searchFilters, page, pageSize)
     );
 };
 export const listComputePlanAggregatetuples = (
     key: string,
+    searchFilters: SearchFilterType[],
     page: number,
     pageSize?: number
 ): AxiosPromise<PaginatedApiResponse<AggregatetupleT>> => {
     return API.authenticatedGet(
         URLS.LIST_AGGREGATETUPLES.replace('__KEY__', key),
-        getApiOptions([], page, pageSize)
+        getApiOptions(searchFilters, page, pageSize)
     );
 };