+++ src/routes/algos/Algos.tsx
@@ -27,7 +27,7 @@ import {
     CreationDateTd,
     CreationDateTh,
 } from '@/components/CreationDateTableCells';
-import OwnerTableFilter from '@/components/OwnerTableFilter';
+import { OwnerTableFilter } from '@/components/NodeTableFilters';
 import PermissionCellContent from '@/components/PermissionCellContent';
 import SearchBar from '@/components/SearchBar';
 import Skeleton from '@/components/Skeleton';